Allow floating point low stock threshold for materials (#12)
Removes the .int() constraint from all material Zod schemas and changes
form inputs from step="1" to step="any" so users can set fractional
thresholds like 1.5 packs.

+++ packages/types/src/baseMaterial/index.ts
@@ -10,7 +10,7 @@ export const baseMaterialSchema = z.object({
     materialCode: z.string().optional(),
     type: z.enum(MaterialType),
     dateAdded: z.string().datetime(),
-    lowStockThreshold: z.number().int().nonnegative().optional(),
+    lowStockThreshold: z.number().nonnegative().optional(),
 });
 
 export type BaseMaterial = z.infer<typeof baseMaterialSchema>;
